+++ tests/fixtures.py
@@ -7,7 +7,7 @@
 # For the full copyright and license information, please view
 # the LICENSE.txt file that was distributed with this source code.
 
-from environ.compat import json
+import json
 
 
 class FakeEnv: